requirements.txt
click==8.1.7
anthropic
pdfplumber
cohere
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
crewai==0.130.0 
crewai-tools==0.47.1
fastapi==0.110.3
openai==1.30.5
pdfplumber
python-dotenv
uvicorn
pydantic